Moved search request classes to starter
