+++ Backend/custom-starter/src/main/java/com/customstarter/model/StateCode.java
@@ -1,4 +1,4 @@
-package com.ms.search.model;
+package com.customstarter.model;
 
 import lombok.Getter;
