+++ Backend/custom-starter/src/main/java/com/customstarter/model/request/SearchRequestPayload.java
@@ -1,7 +1,12 @@
-package com.ms.search.model;
+package com.customstarter.model.request;
 
+import com.customstarter.model.Date;
+import com.customstarter.model.Name;
+import com.customstarter.model.RaceCode;
+import com.customstarter.model.SexCode;
+import com.customstarter.model.StateCode;
+import com.customstarter.utils.ImageIndicatorDeserializer;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.ms.search.utils.ImageIndicatorDeserializer;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
