+++ Backend/custom-starter/src/main/java/com/customstarter/model/schema/combination/ApiCombination.java
@@ -1,6 +1,5 @@
-package com.customstarter.model.form;
+package com.customstarter.model.schema.combination;
 
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -8,9 +7,12 @@ import java.util.ArrayList;
 import java.util.List;
 
 @Getter
-@AllArgsConstructor
 @NoArgsConstructor
 public class ApiCombination {
     private String apiName;
     private List<CombinationFieldReference> fields = new ArrayList<>();
+
+    public ApiCombination(String apiName) {
+        this.apiName = apiName;
+    }
 }
