+++ Backend/custom-starter/src/main/java/com/customstarter/model/schema/combination/CombinationFieldReference.java
@@ -1,4 +1,4 @@
-package com.customstarter.model.form;
+package com.customstarter.model.schema.combination;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
