+++ Backend/custom-starter/src/main/java/com/customstarter/model/schema/input/HtmlInputType.java
@@ -1,4 +1,4 @@
-package com.customstarter.model.form;
+package com.customstarter.model.schema.input;
 
 public enum HtmlInputType {
