+++ Backend/microservices/data/src/main/java/com/ms/data/service/builders/select/ImageIndicatorSelectBuilder.java
@@ -1,8 +1,8 @@
 package com.ms.data.service.builders.select;
 
-import com.customstarter.model.form.input.HtmlInput;
-import com.customstarter.model.form.input.Select;
-import com.customstarter.model.form.input.SelectOption;
+import com.customstarter.model.schema.input.HtmlInput;
+import com.customstarter.model.schema.input.Select;
+import com.customstarter.model.schema.input.SelectOption;
 import com.ms.data.dto.xml.Field;
 import com.ms.data.service.SearchFormBuilderService;
 import com.ms.data.service.builders.HtmlInputBuilderByNameAndType;
