+++ Backend/microservices/data/src/main/java/com/ms/data/service/builders/select/RaceCodeSelectBuilder.java
@@ -1,28 +1,24 @@
 package com.ms.data.service.builders.select;
 
-import com.customstarter.model.form.input.HtmlInput;
-import com.customstarter.model.form.input.Select;
-import com.customstarter.model.form.input.SelectOption;
+import com.customstarter.model.RaceCode;
+import com.customstarter.model.schema.input.HtmlInput;
+import com.customstarter.model.schema.input.Select;
+import com.customstarter.model.schema.input.SelectOption;
 import com.ms.data.dto.xml.Field;
 import com.ms.data.service.SearchFormBuilderService;
 import com.ms.data.service.builders.HtmlInputBuilderByNameAndType;
 import org.springframework.stereotype.Component;
 
+import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Component
 public class RaceCodeSelectBuilder extends HtmlInputBuilderByNameAndType {
 
-    private static final List<SelectOption> OPTIONS = List.of(
-            new SelectOption("", "Unknown"),
-            new SelectOption("W", "White"),
-            new SelectOption("B", "Black or African American"),
-            new SelectOption("A", "Asian"),
-            new SelectOption("L", "Latino or Hispanic"),
-            new SelectOption("I", "American Indian or Alaska Native"),
-            new SelectOption("H", "Native Hawaiian and Other Pacific Islander"),
-            new SelectOption("O", "Other")
-    );
+    private static final List<SelectOption> OPTIONS = Arrays.stream(RaceCode.values())
+            .map(raceCode -> new SelectOption(raceCode.name(), raceCode.getRace()))
+            .collect(Collectors.toList());
 
     @Override
     protected boolean canBuild(String name, String type) {
@@ -32,6 +28,7 @@ public class RaceCodeSelectBuilder extends HtmlInputBuilderByNameAndType {
     @Override
     public HtmlInput build(Field field, SearchFormBuilderService searchFormBuilderService) {
         Select select = new Select(field.getName(), field.getDescription());
+        select.getOptions().add(SelectOption.UNKNOWN);
         select.getOptions().addAll(OPTIONS);
 
         return select;
